requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
plotly>=5.0.0
requests>=2.25.0
